machine: rename sa8775p-ride-sx to qcs9100-ride-sx
Follow Qualcomm Linux downstream and rename sa8775p-ride-sx machine to
qcs9100-ride-sx, including the device-tree files available upstream for
both the sa8775p and qcs9100 variants.

qcs9100 is a variant of sa8775p, and they are fully compatible with
each other.

Firmware, cdt and partition confs updated as part of the same commit as
there is no need to maintain the sa8775p-ride-sx machine override
anymore.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>