conf: add qrb2210-rb1 machine
This is the machine configuration file for the Qualcomm RB1
platform.

There are 2 'hacks' in there.
They are needed because the firmware does not support (yet) EFI
boot. Up until then let's build the corresponding Android boot
image. To flash an RB1, the boot image "boot-qrb2210-rb1-core-kit.img"
needs to be flashed in the 'boot' partition, while the EXT4 image of
the rootfs (e.g. "core-image-base-qrb2210-rb1-core-kit.rootfs.ext4")
should be flashed in 'rootfs' partition. And qbootctl is needed to
inform the Android bootloader that the boot was successful.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>