ci: force build of new/updates recipes
In a PR test build, it's possible to have a new recipe (or updates
with a bbappend) which is not included in any image we build. We could
do a 'world' build but that would be really expensive. Instead this
change tries to catch any updated .bb or .bbappend file, and force a
bitbake build.

If the recipe was already built as part of the image, it should be a
quick no-op. If the recipe was not built, then it's a good thing to
check that it builds!

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>