ci: use KAS_CLONE_DEPTH instead of KAS mirrors
We initially used KAS_REPO_REF_DIR to keep a local copy of the repos
mirrors, to avoid fetching projects from scratch in CI. However, it is
implemented such that the mirrors, once created by kas the *first*
time, are never updated by kas, they are 'read only'.

Some others (such as meta-arm) have implemented a script to update the
repos manually, however KAS has a variable called KAS_CLONE_DEPTH,
which as the documentation says "This is useful in case CI always
starts with empty work directory and this directory is always
discarded after the CI run". Which is exactly what we need here.

Lately, the Yocto git servers have been flaky, which caused KAS to
fail to fetch from upstream, and the KAS ended up falling back to the
old content of the mirrors, silently. e.g. doing CI with old refs for
oe-core, bitbake, .. which is very much undesirable.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>