u-boot: add recipe to build latest mainline plus RB1 patches
This builds a u-boot that can be flashed as boot.img on RB1 and will
load an EFI bootloader from ESP.
An android boot image is generated and deployed based on u-boot.bin and
the matching dtb.

COMPATIBLE_MACHINE is set to 'qcm2290' in anticipation of future
machines using this chainload method.

Signed-off-by: Koen Kooi <koen.kooi@oss.qualcomm.com>