partition-utils: use qcom-ptool
qcom-ptool is a new project which is part of Qualcomm Linux, where we
intend to provide all partition tools to generate GPT images, flashing
artefact, ... These tools were maintained in various places up until
now, and there was a copy of them in meta-qcom for some of them. With
this patch, we removed the local copy from meta-qcom, and fetch all
the tools we need from the new 'upstream' qcom-ptool.

Having these tools in a single, maintained repo (upstream) will make
it easier for other distro to use them (such as our Qualcomm Linux
Debian flavor).

Since qcom-ptool provides both ptool.py and gen_partition.py, the
qcom-gen-partitions-tool recipe is simply removed from here.

This patch is effectively a no-op, and the tools we fetch are (at the
time of this commit) exactly the same.

qcom-ptool also has a copy of each platform partition configuration
file. For now, we are not using them in meta-qcom, and instead we are
keeping our own copy. It's not clear yet, how we will manage the
partition config files over time, over multiple LTS releases.. and
it's not clear that we can have a single partition config file that
will work everywhere. This change can be done later if needed.

The LICENSE md5 change is because of cosmetic changes in the LICENSE
file, there is no change in the license itself.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>