ci: Use kas container for builds
We currently have a custom VM image where we manage build dependencies.
This process is a bit too opaque and makes it hard for maintainers to
manage this project without IT help.

This change moves to the kas container and provides the ability to
install extra build dependencies we may need.

Signed-off-by: Andy Doan <andy@foundries.io>