ci: build-yocto: switch to kas-container
Doing the build with kas-container will satisfy our requirements as it has most of the
necessary tools. This container was created especially to do yocto build and it already
has a good community of users who use it. Its development is quite active and has a
significant group of contributors.

This would also make things easier for our customers and end users as they would only
have to install docker or podman on their machines to replicate what we do on our ci builds.

Inside the kas-container the directory tree is the following:

/build is the bitbake build folder
/repo is the meta-qcom layer
/work is the work directory with all the other layer, bitbake and also the above build dir:
/work
├── bitbake
├── build
├── meta-yocto
└── oe-core
When the KAS_WORK_DIR is not defined the /repo and /work have the same content.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>