ci: yocto-check-layer: switch to kas-container
Adapt the script to be used inside the kas-container.

The KAS build is mounted on /build so we force it anymore to be
outside of the checked out repo.
The KAS_WORK_DIR is mounted on /work.

Running the yocto-check-layer-wrapper fails inside the contianer.
The main task performed by this is to create a clean dir to execute
the real yocto-check-layer, so we can do the same.

The /bin/bash is because the source command is not provided by /bin/sh.

We clone the meta-qcom again in the new builddir from the /repo
to avoid any polution/untracked files that may exist in the repository.
This creates a clean copy of the meta-qcom repo.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>