DONT MERGE: linux-qcom-next: add v6.15 kernel recipe
Add a recipe for the linux-qcom-next 6.15 kernel.

This recipe compiles the qcom-next kernel tree, which includes additional
patches intended for upstream submission on top of the mainline kernel.

Starting with v6.15, this tree will continue to track the latest kernel
releases (e.g., 6.16-rc1, 6.16, 6.17 etc.), with patches being dropped
or rebased as needed. The eventual goal is to upstream all patches.

Signed-off-by: Viswanath Kraleti <viswanath.kraleti@oss.qualcomm.com>