ci: add kas-container-shell helper script
This helper is very similar to the kas-shell-helper but it is intendend
to use the kas-container.

The TOPDIR and KAS_WORK_DIR cannot be changed within the container and
are therefore constant. The SCRIPT path need to be relative to the TOPDIR
also because of the mapping in the container.

Finally the KAS_CONTAINER allows to use non-standard kas-container location
in our ci.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>