workflows: save build_url for each machine separately
Save build_url-<machine name> files as build artifacts. This change
helps to re-run failed builds in PR. Till now the assumption was that
all builds share the same upload URL. In case build jobs fail for
infrastructure reasons and only failed builds are restarted, the URLs
are not the same for all machines. This patch allows the test jobs to
use appropriate URL for each machine separately.

It should be noted that machine names in kas files, build-yocto.yml and
ci/lava directory must be kept in sync for this code to work properly.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@oss.qualcomm.com>