recipes-graphics: Add recipe for qcom-adreno
This patch introduces a new recipe for qcom-adreno,
which fetches Qualcomm Adreno User Mode libraries
from QArtifactory. These libraries include a set of
user-mode binary drivers that support OpenGLES,
Vulkan, and OpenCL APIs.

The recipe builds an ipk package (qcom-adreno) that
enables Adreno GPU support on top of KGSL. It has
dependencies on libdmabufheap, the gbm MSM backend,
and graphicsdlkm.

Once the qcom-adreno overlay ipk is installed and
the system is rebooted, the msm_kgsl driver will be
loaded. To disable the GPU component in the DRM driver,
the no_gpu module parameter is intended to be passed.
However, support for this parameter in the DRM driver
is currently under development and not yet available.

Signed-off-by: Sai Sumanth Sadhula <ssadhula@qti.qualcomm.com>