ci: runs the yocto-check-layer with nodistro and poky
By default our yocto-check-layer uses the oe-core nodisto.
We can specify the kas yml config to be used on the yocto-check-layer
so we will add another job to run the yocto-check-layer using the
Yocto poky distro.

The yocto-check-layer jobs don't take any adavantage of the bitbake sstate-cache
and it will take some time when they runs sequential. So in parallel we reduce to
half of what we would need if they were sequential.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>