ci: naming the download/upload of the kas lockfile
Next patch will introduce new artifact downloads and uploads in the steps.
It is better to start naming that steps of the job so that they are easier
to identify when we look at the logs.
Also rename the artifact to kas-lockfile since it is more obvious
that it represents a kas locking file.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>