ci: base: workaround to temporary disable the libsolv in opkg
The libsolv crashes the opkg in some scenarios not yet fully known.
The proper fix is ​​to disable the libsolv in opkg while looking for
the solution to the problem.

It was observed during debug testing that are still on going that
the issue does not exist when the libsolv is disabled in opkg.

For now it fixes the segfault [1]:
| kernel: opkg[3077180]: segfault at 7fd957c88b28 ip 00007fd957e162f2 sp 00007fffb6068490 error 4 in libsolv.so.1[142f2,7fd957e0e000+7c000] likely on CPU 4 (core 8, socket 0)

[1] https://bugzilla.yoctoproject.org/show_bug.cgi?id=16010

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>