ci: base: fix oe-core SPDX sstate-cache artifacts
If we have changes on SPDX_LICENSES content we ended up building invalid sstate-cache archives.
The default value for the SPDX_LICENSES is the file meta/files/spdx-licenses.json but this file
don't use the bitbake fetcher and because of this their checksum is not validated.
So we need to add this file to the build dependency chain of the SPDX.

Signed-off-by: Jose Quaresma <jose.quaresma@oss.qualcomm.com>