CI: Add KAS yaml file to build open boot firmware stack
Currently the open boot firmware stack has only been enabled on RB3Gen2 platform
which can be built using following:

$ kas build ci/qcs6490-rb3gen2-core-kit.yml:ci/open-boot-firmware.yml

And currently the boot stack includes TF-A, OP-TEE and U-Boot. In future the
plan is to enable upstream edk2 as well. Along with that the next target for open
boot firmware is Lemans based IoT EVK platform.

Right now the build generates 2 firmware payloads as bl2.elf (unsigned) and fip.elf
(test signed using qtestsign). It is required to sign bl2.elf with QTI signature
using sectools but in future the plan is to drop QTI signature requirement with an
updated release of XBL/XBL-SEC.

Once signing is done, one need to update following binaries in qcomflash tarball
and then proceed with QDL flashing:
- tz.mbn -> bl2.mbn
- uefi.elf -> fip.elf

Signed-off-by: Sumit Garg <sumit.garg@oss.qualcomm.com>