resize-rootfs_1.0.bb: add recipe to install rootfs resize script
To be able to flash across various boards, the root filesystem is generated
with a minimal size. This `resize-rootfs.sh` script, automatically resizes
the rootfs on first boot to utilize the full available disk space.

The script is installed via a systemd service (`resize-rootfs.service`) that
runs on first boot and disables itself afterward.

Signed-off-by: Viswanath Kraleti <viswanath.kraleti@oss.qualcomm.com>