qcom-common: use standard settings for EFI options
Changes to EFI_PREFIX, EFI_DTB_DIR and KERNEL_DTBDEST should not be set
by the BSP layer and should be controlled entirely by the distro.

This was done in order to generate esp-qcom-image with the right file
layout, but that should be done at the image level and not at the
package level.

Signed-off-by: Ricardo Salveti <ricardo.salveti@oss.qualcomm.com>