image_types_qcom: fail on incorrect QCOM_PARTITION_FILES_SUBDIR
If QCOM_PARTITION_FILES_SUBDIR is invalid, current create_qcomflash_pkg
code will happily ignore missing directory and create the incomplete
flashing directory (as happened for the QCS615 Ride platform). In order
to prevent this from happening in future, remove the find calls and
refer the necessary files using the wildcards. If the dir is not
correct, wildcard matching will not be expanded and thus install command
will fail with the 'No such file or directory' error thus not letting
the incorrect configuration to build successfully.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>