linux-qcom-next: enable I2C_QCOM_GENI as built-in
On Lemans and Monaco EVK platforms the boot-from-SD-card feature requires
an IO expander to be enabled at an early stage to detect the SD card. This
IO expander is connected over I2C.

Currently the I2C driver is built as a loadable module which resides on the
SD card along with other modules. This creates a circular dependency where
to detect the SD card the IO expander must be initialized the IO expander
depends on the I2C driver and the I2C driver is on the SD card which cannot
be accessed until detection succeeds.

To break this dependency enable I2C_QCOM_GENI as built-in so that the
driver is available during early boot.

Signed-off-by: Viken Dadhaniya <viken.dadhaniya@oss.qualcomm.com>